requirements.txt
Django==5.1
djangorestframework==3.15.2
django-environ==0.11.2
django-cors-headers==4.4.0
djoser==2.2.3
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.2
whitenoise==6.7.0
psycopg2==2.9.9
argon2-cffi==23.1.0
pillow==10.4.0
vonage==3.17.1